+++ src/hario_core/models/__init__.py
@@ -1,0 +1,8 @@
+from .extensions.chrome_devtools import DevToolsEntry
+from .har_1_2 import Entry, HarLog
+
+__all__ = [
+    "Entry",
+    "HarLog",
+    "DevToolsEntry",
+]
