+++ src/hario_core/transform/strategies.py
@@ -2,8 +2,12 @@ from abc import ABC, abstractmethod
 from concurrent.futures import ProcessPoolExecutor, ThreadPoolExecutor, as_completed
 from typing import Any, Dict, List, Optional
 
-from hario_core.interfaces import Transformer
-from hario_core.worker import init_worker, process_batch
+from hario_core.transform.interfaces import Transformer
+
+from hario_core.transform.worker import (
+    init_worker,
+    process_batch,
+)
 
 
 class ProcessingStrategy(ABC):
